Implement exercise 1.11 - recursive vs iterative
The iterative version currently calculates 2 iterations beyond what is
actually required of it. It could be optimized by starting from
iteration 3 and returning `c` instead of `a` in the end condition, but
this would require cases where `n < 3` to be handled separately with a
top level cond, for example.

The recursive version could of course also be optimized by using
memoization, for example.