E1.29 - calculating integrals with Simpson's rule
Running this with (simpsons-integral cube 0 1 n) with n being any even
number, it currently returns 1/4 - an exact rational number. This is
unexpected, because this was supposed to be an approximation. Changing
the 3 to be 3.0 and wrapping n in (exact->inexact n) in the division for
h changes the result to be a float and look more like an approximation,
but I'm not sure if that is to be desired here. Doing these calculations
by hand for n=2 shows that the result can be found using just rational
numbers and comes out to be 1/4, so maybe it's best to let Scheme do
it's thing and not force it to use floats.