E1.30,1.31 - product functions + pi approximation
Otherwise pretty straight forward, but figuring out the term function
for the pi approximation was a bit tricky.

Once again the question of exact/inexact arose. Didn't want to force
inexact calculations, but wanted to see the result as a decimal number
to confirm it's proximity to 3.14... Without using exact->inexact on the
result, it was a rational number with very big numberator and
denominator.