E1.37a - Approximating golden ratio with continued fractions
I know there's probably better ways of checking when something is
"accurate to 4 decimal places" (e.g. comparing the delta to the smaller
of the differences between the current value and closest values with
different 4th decimal places), but for sake of simplicity I wrote this to
check the delta to be within the specified tolerance. For a tolerance of
0.0001 this took 11 steps and resulted in 1.6180555555555558 (which is
in fact accurate to the 4th decimal, if the following decimals are cut
of (floored) instead of rounded up as would be correct for the 5s here).