E1.38,1.39 - Approximating e and tan
The tan function kept giving me weird results for the longest time and I
couldn't figure out what's amiss. Turned out I'd gotten the n and d
functions mixed up. It's probably too late for doing this.

This is the output of the test run I used for the tan approximation:
```
> (approximate-to-precision (lambda (k) (tan-cf-approx 2.0 k)) 0.00001)
'(8 -2.185039871780394)
```