E1.45 - Nth root calculation using automatic fixed-point damping
After playing around with the numbers a little, it looked like the
function had to be damped log2(n) times. This approach seemed to work
fine when I tested values of n up to 128. I'm not sure, however, if a
less aggressive (i.e. one that uses less damping) function wouldn't
work as well. Also the number 2 as the logarithmic base seems a bit
arbitrary.

Added a safeguard to `fixed-point` that makes it give up after 1000
guesses. This isn't enough for larger numbers, because `fixed-point`
uses an absolute tolerance value for it's `close-enough?` check, but
seemed good enough for the smaller numbers I tested with.

I was kind of lazy and so re-defined `repeated` to essentially repeat
the provided amount of times rounded down. So it doesn't expect an
actual integer anymore.