E2.1 - Normalizing rational number constructor
Thought of a couple of different options for doing this, but this seemed
the most elegant. One other option that crossed my mind was to divide
both `n` and `d` with their absolute values and multiply the results to
get the sign for the numerator. The _sign_ would then be multiplied with
the absolute value of the provided (or normalized) numerator and the
absolute value of the provided (or normalized) denominator would be
used.