E2.2,2.3 - Segment and rectangle representations
Thought about normalizing segments so that, for example, start-segment
would always return the point whose x is smaller. But wasn't sure if
that's necessary and so didn't do it. However, later with the rectangles
it might make sense to do something like this, otherwise the semantics
of the "height", for example, might be vague. However, if "height" is
defined as "a segment of with the provided length that's in rotate 90
degrees counterclockwise to the provided segment's angle", then there
shouldn't be any ambiguity left.

Had some trouble figuring out ways to represent rectangles. There's many
possible ways, but most also allow storing non-rectangles. For example,
a simple solution would be to store 4 points, and just take the user's
word for it that the points form a rectangle. However, any shape with 4
corners could be represented in that fashion and so the functions are
not guaranteed to always work as expected.

For some time I had the idea that a rectangle could be represented by
it's diagonal, and it took me a while to prove myself wrong. This final
result is OK, I think.