E2.4-2.6 - Representing data through procedures
I needed some sort of a function to confirm whether the numbers I was
creating were correct. Initially created a function that printed as many
dots onto the screen as was the number's value. Then realized I could
actually print the real number, by changing the dot-printing function to
just increment the value provided to it and seeding it with 0. Then
realized that there's no need to do the printing, when I can just find
the number and have the procedure (now named `eval`) return the value.

I don't _really_ understand why my implementation of `plus` works, but I
had a hunch and tried and it does seem to work. I have the beginnings of
an understanding but can't claim to actually fully understand.