E2.11 - Optimize interval multiplication with sign testing
The idea is that an interval can be in one of three mutually exclusive
states: 1) none of its endpoints are positive, 2) it's upper bound is
positive, while the lower isn't, and 3) both of its bounds are positive.
This means that a pair of intervals can be in a combination of those
three states, which 9 separate cases were hinted at by the exercise and
are implemented below.

I feel like there must be a simpler way to represent this logic, but
can't think of anything right away that wouldn't require my to implement
a special form of data and an interpreter for it.

Confirming that this is actually correct is a pain without a property
testing framework and because I don't have anything like that set up
here, I am not fully convinced that the specifics are correct. A quick
sanity check seemed to pass however, so it works in at least some of the
cases.