E2.12 - Create a percentage-based interval constructor
make-center-percent can provide the bounds to the base constructor in
wrong order (when center is negative), but that's not an issue, because
that constructor already checks for these cases and swaps the bounds.

I changed the width-interval implementation, because it's effectively
the same, but it looked curious (albeit non-obvious) being implemented
this way.