E2.21-2.23 - List mapping, traversing
Straight forward stuff.

Answer for E2.22: Why are the results reversed? Heh, because that's the
algorithm to reverse a linked list. In all seriousness, though, it
starts from making a list of the first element and nil, and then the
second element and the list formed in the first step and so on. All this
with a transforming operation applied to all the elements.

As for why the fix doesn't work, well because by how lists are defined,
the first element of the pair should be a single element, not a list.
This implementation essentially reverses the convention, by setting the
"tail" of the list as the first element of the pair. But this still
doesn't re-reverse the list, because one would still have to traverse
the entire structure to reach the last element.