E2.29d - Changing underlying structure
As the diff shows, only the selectors had to change, because nowhere
does the rest of the code rely on the internal structure of these
objects.

There are the `pair?` checks to know whether a `branch-structure` is a
weight or another mobile, and I guess that could be abstracted away with

```scm
(define (weight-branch? branch)
  (not (pair? (branch-structure branch))))
```

Not sure if that's worth it, though.