E2.33-2.39 - Foldin' n' mappin'
Exercise 2.38 asks a property that an operation should satisfy for
`fold-left` and `fold-right` to yield the same result for all sequences.
The operation must be both associative and commutative.

I think it's possible to relax the constraint to just associative,
however, when dealing with non empty sequences and if the initial value
is thought of as just another value in the sequence. Except that for
`fold-right` it is appended to the sequence, while for `fold-left` it is
prepended to the sequence. E.g.:

```scm
(= (fold-left op x (cons y rest))
  (fold-right op y (append rest (list x))))
```

Reading the docs, somewhat similar things apply to  `reduce` and
`reduce-right` - they ignore the initial value if there's at least a
single element in the provided sequence.