E2.42,2.43 - Eight (or N) queens puzzle
The `safe?` check works by projecting 3 lines to the right of the queen
being tested. One going diagonally up, one going in a straight line to
right and one going diagonally down. For every column after the first
one, it checks whether there's a queen on any of the 3 lines. If yes, we
know that the new queen would be checked and isn't "safe".

Initially thought of storing a list of boolean values for the column,
but then realized, that all I need are the position of the queens, which
can simply be stored as numbers. The `column-number` value turns out
not to be useful at all, but I kept it because seeing the pair in an
output makes more sense (is easier to understand) than just seeing the
number, although they're equivalent in utility for this problem.

Used `apply` and the way `map` works with multiple lists to make
serialization work. It works sort of as a matrix transposer.

As for E2.43, the recursive call to `queen-cols` is called `board-size`
times in our implementation, but in the slow implementation it will be
called `board-size` to the power of `board-size` times. So if it the
program in this commit takes time T to solve a problem, then Louis'
program would probably take roughly `T*n^(n-1)`, where `n` is the
`board-size`.

I guess computers were slower back then. A 6x6 board doesn't really take
all that long with current computing power.