Implement flipping to support square-limit paintings
Kind of feel sad if I have to start removing all of this when continuing
with the exercises from the book which seem to be veering in the
direction of painters being functions that take frames as inputs. Not
sure how easy it is to reconcile that with my current implementation of
painters being data.

In any case, this now allows me to get this output when displaying
`(paint (square-limit painter 1))`:

```
X--XXXXXXXXXX--X
|  ||||||||||  |
|  ||||||||||  |
X--XXXXXXXXXX--X
X--XX--XX--XX--X
X--X|  ||  |X--X
X--X|  ||  |X--X
X--XX--XX--XX--X
X--XX--XX--XX--X
X--X|  ||  |X--X
X--X|  ||  |X--X
X--XX--XX--XX--X
X--XXXXXXXXXX--X
|  ||||||||||  |
|  ||||||||||  |
X--XXXXXXXXXX--X
```