Actually draw onto the screen
Had to change the interface of painters to also accept the
`draw-line-canvas` function to make this work.

Ideally the frame would get its size information from the canvas,
instead of the size defined in the `on-screen` method having to somewhat
match that defined for `a-frame`, but can look at that later.

Had to install GUI libraries for this to work. I used the following
command for this:
```
raco pkg install --auto gui
```
Although, I think installing `gui-lib` and maybe `draw-lib` would've
been enough.