E2.52 - Add a smile to the wave painter
I did do the other 2nd part of the exercise as well, but don't want to
commit it, because then the square-limit result looks worse. I want to
show off the square-limit result later so I'm keeping the best version.
The 3rd part of the exercise is also pretty trivial and because I don't
actually have the `square-of-four` version of square-limit included in
my code I decided it's not worth adding it now either.