E2.60 - Set as ordered list with duplicates
The complexity of `adjoin-set` changed from O(n) to O(1). The complexity of
`union-set` changed from O(m*n) to O(n), where n is the size of the first set
(actually it's underlying list's length) and m is the size of the second set.

The complexity of `element-of-set?` and `intersection-set` ostensibly remained
the same, while actually they get worse. Their big O notation complexity didn't
change, but the value of n is generally going to be larger, because it includes
duplicate values as well.

This implementation is good in cases where insertions into the set need to be
very fast and memory exhaustion will not be a problem. For example in
situations where data collection into sets needs to be very fast, but the list
can later be deduped. When there's no more time pressure.