E2.63 - Comparing 2 tree->list functions
As exemplified by the last expression in the file evaluating to `#t`, the two
tree->list functions create the same output. The function `tree->list-1` is
slower, because in addition to everything that `tree->list-2` does, it also
uses `append` which is linear in complexity of the first value provided to it.
That is, the size of the list created from the trees left branch.