E2.64,2.65 - Set union and intersection in linear time
How does `partial-tree` work? It works by dividing the provided elements (or
a subset of them) into 3 groups, the middle element, the left-side elements and
the right-side elements. Left and right side usually get even number of
elements, unless an uneven number of elements has to be divided by the two
sides, in which case the right side is preferred. It takes these elements and
recursively builds the tree, recursing down to both sides. It does so in linear
time, so O(n).

Because `tree->list-2`, `list->tree`, `union-ordered-list`, and
`intersection-ordered-list` ordered list are all linear in complexity, their
combination, as done in `union-set` and `intersection-set`, is also linear, as
requested by exercise 2.65.