Write test with unless + error
This way I can keep these tests within the files as the files keep evolving,
and have the benefit of the test failures stopping the entire execution. Right
now the tests would always have to be last elements in the file and the result
can be seen when the file is loaded or test is evaluated directly.