Fix or and improve context quality for both and+or
Apparently the nice commonality found between `and` and `or` in 4d38511 ("Add
support for not procedure", 2017-09-17) wasn't actually a commonality at all
, but rather a misunderstanding of the previous implementation. This caused
a bug in or, which is fixed here.

I like how well this implementation now encodes a big idea behind this
approach. Namely, for both `and` and `or`, combine both assertions if they're
both successes or failures. This makes sure that whatever the final outcome, we
have the maximal amount of context for the "cause" of the result. And then the
difference between `and` and `or` is also very clear in the implementation - if
the two assertions aren't of the same type, then for `and` failures take
precedence and for `or` the successes take precedence.