Support pending assertions
If an assert only includes the description, it will be considered a pending
assertion and a warning will be printed every time the assertion is executed.

Separate some tests into a different file, because otherwise these tests would
start spamming the stdout every time the testing library itself is loaded.