E2.73(1,2) - Differentiate by dispatching on type
As an answer to part 1 of the exercise, we can't dispatch based on type to
numbers and variables, because they're primitive values and so don't have
a type tag. In order for them to have type tags, they would have to always be
constructed through some constructors provided by us. But this makes for a bad
interface, if the user of your function can't use simple numbers, but has to
use some `create-number` function instead.