E2.74(3,4) - Find employee record from a list of files
The tests don't really show it, but this would also work with files of
different types (coming from different divisions).

As an answer to part 4, if there's a new division, then that division has to
create its own installation package, while other divisions and the HQ don't
have to change anything.