E-2.75,2.76 - Complex numbers with message passing
For E2.76, in a system where new types must often be added, message passing
might make more sense and in a system where new operations must often be added,
dispatching on type makes more sense. The reasoning being that with message
passing one has to always carry all the operations along with the data. This
might be unwieldy when the number of operations is big. Separating the
operations from the data and dispatching on type, by looking up the operation
from a table avoids this issue.

Generic operations with explicit dispatch might be a valid approach if the
amount of types remains constant throughout the evolution of the program, while
the amount of procedures keeps growing. This is because this approach then
allows you to define the operation for all different types in a single location
when adding that operation.