E2.77 - Dealing with multiple levels of tagging
As explained in the book, the complex number has two levels of tags now. Its
actual value is something like `'(complex (polar <val>))`. When we define these
operations in the "complex" package, then these operations will be called
within that package with the _contents_ of the value, which is another tagged
value, namely: `'(polar <val>)`. And when we call apply-generic with this
internal tagged value, then it finally reaches the correct package that
actually knows what to do with the value - either polar or rectangular.