E2.80 - Add zero? check for for numbers
I wasn't sure if that's how you test for "zero" for complex numbers, but some
googling reveals that it should indeed be correct. WolframAlpha wiki says that
a complex number is nonzero if either its real part or imaginary part is
nonzero. Seeing as the magnitude is the sum of the square of both of these
parts, the magnitude can only be zero if both of them are zero. Therefore
checking for the value of the magnitude is enough.

Had to define `square` because it was used for complex number conversion from
one type to another, but wasn't defined yet.