E2.81 - Avoid redundant coercions
As an answer to part one, if Louis Reasoner's approach were implemented
and types were automatically coerced to themselves, then calling the
procedure `exp` included in this commit would result in an infinite loop.
That is because the values would be coerced to the types they already are
and then apply-generic would be called with the coerced values, which
would of course repeat the same approach.

As an answer to part two, `apply-generic` already works fine as it is. The
suggestion provided in part 3 to avoid trying to coerce types to
themselves avoids some unnecessary computation, but doesn't change
behavior. This is because coercion is only tried after it is found that
the generic procedure does not exist for the provided types. Because the
error is the same for the cases where coercion isn't found and where the
procedure isn't found it doesn't matter if it tries to coerce the types to
themselves or not.