E2.82 - Coercing arbitrary number of arguments
This fails in the same cases as the two argument version, as described in the
book. For example, when none of the arguments in the provided test case are of
type 'complex-number, then the current algorithm wouldn't try to coerce the
arguments to that type, even if it could. This pretty much generalizes the
faulty two-argument approach to more arguments, without fixing its issues.