E2.84 - Use raising in apply-generic
Not exactly pretty, but it demonstrates the idea pretty well, I think.

This would be more elegant, if the type information of all installed generic
operations also included the type of the return value. In that case, we could
just walk through the meta data of the generic "raise" operation and see if
with multiple application we could get to the desired type, without actually
having to do the raising as done now.