E2.85 - Add projection to lower types
Removed the "real" type, because it's plenty of work without it already.

This could've been much simpler if proper packages were installed from the
beginning for the different types, as is done for some earlier exercises. Also,
implicit tagging of primitive numbers would've simplify this quite a bit as
well. But once again, this is enough to demonstrate the effects of the type
system.